+++ src/main/kotlin/moirai/semantics/core/CostExpression.kt
@@ -30,6 +30,9 @@ internal fun sortCanonical(costExpressions: List<CostExpression>): List<CostExpr
     res.addAll(costExpressions.filterIsInstance<ProductCostExpression>())
     res.addAll(costExpressions.filterIsInstance<SumCostExpression>())
     res.addAll(costExpressions.filterIsInstance<ParameterHashCodeCost>().sortedBy { it.typeParameter.qualifiedName })
+    res.addAll(costExpressions.filterIsInstance<InstantiationHashCodeCost>().sortedBy {
+        (it.instantiation.substitutionChain.terminus as ParameterHashCodeCost).typeParameter.qualifiedName
+    })
     res.addAll(costExpressions.filterIsInstance<FinTypeParameter>().sortedBy { it.qualifiedName })
     res.addAll(costExpressions.filterIsInstance<Fin>().sortedBy { it.magnitude })
     res.addAll(costExpressions.filterIsInstance<ConstantFin>())
@@ -39,6 +42,82 @@ internal fun sortCanonical(costExpressions: List<CostExpression>): List<CostExpr
     }
 
     return res
+}
+
+internal fun costExpressionFromAnyType(type: Type): CostExpression {
+    fun fromToString(member: Symbol): CostExpression {
+        return when (member) {
+            is GroundMemberPluginSymbol -> {
+                member.costExpression
+            }
+
+            is ParameterizedMemberPluginSymbol -> {
+                member.costExpression
+            }
+
+            else -> langThrow(TypeSystemBug)
+        }
+    }
+
+    return when (type) {
+        is ConstantFin -> type
+        is Fin -> type
+        is FinTypeParameter -> type
+        is MaxCostExpression -> MaxCostExpression(type.children.map { costExpressionFromAnyType(it) })
+        is ProductCostExpression -> ProductCostExpression(type.children.map { costExpressionFromAnyType(it) })
+        is SumCostExpression -> SumCostExpression(type.children.map { costExpressionFromAnyType(it) })
+
+        is PlatformObjectType -> {
+            fromToString(type.fetchHere(Identifier(NotInSource, StringMethods.ToString.idStr)))
+        }
+
+        is BasicType -> {
+            fromToString(type.fetchHere(Identifier(NotInSource, StringMethods.ToString.idStr)))
+        }
+
+        is TypeInstantiation -> {
+            when (val parameterizedType = type.substitutionChain.terminus) {
+                is ParameterizedBasicType -> {
+                    fromToString(parameterizedType.fetchHere(Identifier(NotInSource, StringMethods.ToString.idStr)))
+                }
+
+                is ParameterHashCodeCost -> {
+                    costExpressionFromAnyType(type.substitutionChain.replay(parameterizedType.typeParameter))
+                }
+
+                is ParameterizedRecordType -> {
+                    SumCostExpression(parameterizedType.fields.map {
+                        costExpressionFromAnyType(type.substitutionChain.replay(it.ofTypeSymbol))
+                    })
+                }
+
+                is PlatformSumRecordType -> {
+                    SumCostExpression(parameterizedType.fields.map {
+                        costExpressionFromAnyType(type.substitutionChain.replay(it.ofTypeSymbol))
+                    })
+                }
+
+                is PlatformSumType -> {
+                    SumCostExpression(parameterizedType.memberTypes.map {
+                        costExpressionFromAnyType(
+                            type.substitutionChain.replay(
+                                when (it) {
+                                    is PlatformSumObjectType -> it
+                                    is PlatformSumRecordType -> it
+                                }
+                            )
+                        )
+                    })
+                }
+            }
+        }
+
+        is GroundRecordType -> SumCostExpression(type.fields.map { costExpressionFromAnyType(it.ofTypeSymbol) })
+        is ObjectType -> Fin(type.identifier.name.length.toLong())
+        is PlatformSumObjectType -> Fin(type.identifier.name.length.toLong())
+
+        else -> langThrow(TypeSystemBug)
+    }
 }
 
 internal class EvalCostExpressionVisitor(val architecture: Architecture): CostExpressionVisitor<Long> {
@@ -57,7 +136,9 @@ internal class EvalCostExpressionVisitor(val architecture: Architecture): CostEx
     }
 
     override fun visit(costExpression: InstantiationHashCodeCost): Long {
-        langThrow(CalculateCostFailed)
+        val tp = costExpression.instantiation.substitutionChain.terminus.typeParams.first()
+        val ce = costExpression.instantiation.substitutionChain.replay(tp)
+        return costExpressionFromAnyType(ce).accept(this)
     }
 
     override fun visit(costExpression: Fin): Long {
@@ -145,7 +226,7 @@ internal object CanEvalCostExpressionVisitor: CostExpressionVisitor<Boolean> {
     }
 
     override fun visit(costExpression: InstantiationHashCodeCost): Boolean {
-        return false
+        return true
     }
 
     override fun visit(costExpression: Fin): Boolean {
