+++ src/main/kotlin/moirai/semantics/infer/Substitution.kt
@@ -75,85 +75,6 @@ internal class Substitution(
             else -> type
         }
 
-    private fun costExpressionFromAnyType(type: Type): CostExpression {
-        fun fromToString(member: Symbol): CostExpression {
-            return when (member) {
-                is GroundMemberPluginSymbol -> {
-                    member.costExpression
-                }
-
-                is ParameterizedMemberPluginSymbol -> {
-                    member.costExpression
-                }
-
-                else -> langThrow(TypeSystemBug)
-            }
-        }
-
-        return when (type) {
-            is ConstantFin -> type
-            is Fin -> type
-            is FinTypeParameter -> type
-            is MaxCostExpression -> MaxCostExpression(type.children.map { costExpressionFromAnyType(it) })
-            is ProductCostExpression -> ProductCostExpression(type.children.map { costExpressionFromAnyType(it) })
-            is SumCostExpression -> SumCostExpression(type.children.map { costExpressionFromAnyType(it) })
-            is ParameterHashCodeCost -> type
-            is InstantiationHashCodeCost -> costExpressionFromAnyType(type.instantiation)
-
-            is PlatformObjectType -> {
-                fromToString(type.fetchHere(Identifier(NotInSource, StringMethods.ToString.idStr)))
-            }
-
-            is BasicType -> {
-                fromToString(type.fetchHere(Identifier(NotInSource, StringMethods.ToString.idStr)))
-            }
-
-            is TypeInstantiation -> {
-                when (val parameterizedType = type.substitutionChain.terminus) {
-                    is ParameterizedBasicType -> {
-                        fromToString(parameterizedType.fetchHere(Identifier(NotInSource, StringMethods.ToString.idStr)))
-                    }
-
-                    is ParameterHashCodeCost -> {
-                        costExpressionFromAnyType(type.substitutionChain.replay(parameterizedType.typeParameter))
-                    }
-
-                    is ParameterizedRecordType -> {
-                        SumCostExpression(parameterizedType.fields.map {
-                            costExpressionFromAnyType(type.substitutionChain.replay(it.ofTypeSymbol))
-                        })
-                    }
-
-                    is PlatformSumRecordType -> {
-                        SumCostExpression(parameterizedType.fields.map {
-                            costExpressionFromAnyType(type.substitutionChain.replay(it.ofTypeSymbol))
-                        })
-                    }
-
-                    is PlatformSumType -> {
-                        SumCostExpression(parameterizedType.memberTypes.map {
-                            costExpressionFromAnyType(
-                                type.substitutionChain.replay(
-                                    when (it) {
-                                        is PlatformSumObjectType -> it
-                                        is PlatformSumRecordType -> it
-                                    }
-                                )
-                            )
-                        })
-                    }
-                }
-            }
-
-            is GroundRecordType -> SumCostExpression(type.fields.map { costExpressionFromAnyType(it.ofTypeSymbol) })
-            is ObjectType -> Fin(type.identifier.name.length.toLong())
-            is PlatformSumObjectType -> Fin(type.identifier.name.length.toLong())
-            is StandardTypeParameter -> ParameterHashCodeCost(type)
-
-            else -> langThrow(TypeSystemBug)
-        }
-    }
-
     fun applyCost(costExpression: CostExpression): CostExpression {
         return when (costExpression) {
             is FinTypeParameter -> {
@@ -175,15 +96,17 @@ internal class Substitution(
             is MaxCostExpression -> MaxCostExpression(costExpression.children.map { applyCost(it) })
             is ParameterHashCodeCost -> {
                 val res = if (solutions.containsKey(costExpression)) {
-                    val solution = solutions[costExpression]!!
-                    costExpressionFromAnyType(solution)
+                    when (val solution = solutions[costExpression]!!) {
+                        is StandardTypeParameter -> ParameterHashCodeCost(solution)
+                        else -> costExpressionFromAnyType(solution)
+                    }
                 } else {
                     costExpression
                 }
                 res
             }
 
-            is InstantiationHashCodeCost -> costExpressionFromAnyType(apply(costExpression.instantiation))
+            is InstantiationHashCodeCost -> InstantiationHashCodeCost(apply(costExpression.instantiation))
         }
     }
 }
